refactor move cost tables
